requirements.txt
amqp==2.4.2
appdirs==1.4.3
asn1crypto==0.24.0
astroid==1.6.5
atomicwrites==1.3.0
attrs==21.2.0
autopep8==1.4.3
backports.csv==1.0.7
backports.functools-lru-cache==1.5
bcrypt==3.1.6
billiard==3.6.0.0
bugsnag==3.6.1
bunch==1.0.1
cached-property==1.5.1
celery==4.3.0
certifi==2018.11.29
cffi==1.14.6
chardet==3.0.4
configparser==3.5.0
contextlib2==0.5.5
crypto==1.4.1
cryptography==2.8
customerio==1.2
defusedxml==0.7.1
dnspython==1.16.0
elasticsearch==7.13.3
enum34==1.1.6
et-xmlfile==1.0.1
falcon==3.0.1
funcsigs==1.0.2
futures==3.1.1
googlemaps==4.4.7
gunicorn==19.9.0
idna==2.8
importlib-metadata==4.6.1
inflection==0.3.1
iniconfig==1.1.1
ipaddress==1.0.22
isodate==0.6.0
isort==4.3.4
jdcal==1.4.1
jsonschema==3.2.0
kombu==4.5.0
lazy-object-proxy==1.3.1
lxml==4.2.5
marshmallow==3.12.2
mccabe==0.6.1
more-itertools==5.0.0
munch==2.5.0
Naked==0.1.31
numpy>=1.19.1
odfpy==1.4.0
openpyxl==2.6.2
packaging==21.0
pathlib2==2.3.3
Pillow==8.3.1
pluggy==0.13.1
protobuf==3.6.1
py==1.10.0
pycodestyle==2.4.0
pycountry==18.12.8
pycparser==2.19
pycryptodome==3.9.9
PyJWT==1.7.1
pylint==1.9.4
pymodm==0.4.2
pymongo==3.7.2
pyotp==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
pytest-order==1.0.0
pytest-ordering==0.6
python-dateutil==2.8.0
python-dotenv==0.10.1
python-mimeparse==1.6.0
pytz==2018.7
PyYAML==5.1.1
redis==3.2.1
reportlab==3.5.68
requests-toolbelt==0.8.0
scandir==1.10.0
schema==0.6.8
-e git+https://sendbox-machine-user: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@github.com/sendbox-software-inc/core.git@headless#egg=Sendbox_Core
singledispatch==3.4.0.3
six==1.12.0
smile-id-core==0.0.13
toml==0.10.2
ua-parser==0.8.0
ujson==1.35
Unidecode==1.0.23
urllib3==1.24.1
user-agents==1.1.0
validator-collection==1.5.0
vine==1.3.0
wrapt==1.10.11
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
zeep==3.2.0
zipp==3.5.0
transitions==0.6.9
pandas~=0.24.2
phonenumbers~=8.12.34
setuptools~=57.0.0
faust~=1.10.4
PyPDF2==1.26.0
aiohttp~=3.8.1